move to dedicated cyto_utils sqlite path with decomposed functionality
preparatory for conversion work and shared functionality
